requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
connexion==2.7.0
Flask==1.1.2
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
openapi-spec-validator==0.2.8
pycodestyle==2.6.0
Pyphen==0.9.5
pyrsistent==0.16.0
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
textstat==0.6.2
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
